Add Dockerfile to build and serve Angular app
This commit includes a Dockerfile that builds the Angular project in a Node.js environment and serves it using an http-server. It structures the creation of the Docker image in two stages: first, it builds the application with the Angular CLI and then, on a clean Node.js environment, it hosts the built application using a simple HTTP server.